undo api change from previous commit
The previous commit made it impossible to override supply the gpg
key via the global `schleuder::adminkeys_path` search directory.
The reason is that the api was such that a missing `admin_publickey`
signified that the key should be searched there.

To stay compatible with this api we must check if this globally
define keyfile exists for a particular admin and only fall back
to wkd if it doesn't.